+++ src/components/Header/components/AuthButtons/index.js
@@ -1,0 +1,1 @@
+export { AuthButtons } from './AuthButtons';
