+++ src/components/Header/components/Logo/Logo.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import styles from '../../Header.module.css';
+import { Link } from 'react-router-dom';
+import logo from '../../../../assets/upTalent.png';
+
+export const Logo = () => {
+	return (
+		<Link to='/home' className={styles.logo}>
+			<img src={logo} alt='UpTalent Logo' />
+			UPTALENT
+		</Link>
+	);
+};
