+++ src/components/Header/components/Logo/index.js
@@ -1,0 +1,1 @@
+import { Logo } from './Logo';
