+++ src/components/Home/Home.module.css
@@ -1,14 +1,14 @@
 .container {
 	margin: 0 auto;
 	max-width: 1300px;
-    height: 100vh;
+	height: 100vh;
 }
 
 .water {
 	margin-top: 100px;
 	display: flex;
-    gap: 50px;
-    height: 65%;
+	gap: 50px;
+	height: 65%;
 }
 
 .waterHeader {
@@ -51,48 +51,48 @@
 	font-weight: 400 !important;
 	font-size: 20px !important;
 	line-height: 24px !important;
-    letter-spacing: 0.1em;
-    position: relative;
-    overflow: hidden;
-    z-index: 1;
-    transition: 0.5s;
+	letter-spacing: 0.1em;
+	position: relative;
+	overflow: hidden;
+	z-index: 1;
+	transition: 0.5s;
 }
 
 .getstarted span {
-    position: absolute;
-    width: 25%;
-    height: 100%;
-    background-color: var(--primary);
-    transform: translateY(150%);
-    border-radius: 50%;
-    left: calc((var(--n) - 1) * 25%);
-    transition: 0.5s;
-    transition-delay: calc((var(--n) - 1) * 0.1s);
-    z-index: -1;
+	position: absolute;
+	width: 25%;
+	height: 100%;
+	background-color: var(--primary);
+	transform: translateY(150%);
+	border-radius: 50%;
+	left: calc((var(--n) - 1) * 25%);
+	transition: 0.5s;
+	transition-delay: calc((var(--n) - 1) * 0.1s);
+	z-index: -1;
 }
 
 .getstarted:hover {
-    color: black;
+	color: black;
 }
 
 .getstarted:hover span {
-    transform: translateY(0) scale(2);
+	transform: translateY(0) scale(2);
 }
 
 .getstarted span:nth-child(1) {
-    --n: 1;
+	--n: 1;
 }
 
 .getstarted span:nth-child(2) {
-    --n: 2;
+	--n: 2;
 }
 
 .getstarted span:nth-child(3) {
-    --n: 3;
+	--n: 3;
 }
 
 .getstarted span:nth-child(4) {
-    --n: 4;
+	--n: 4;
 }
 
 .waterRightside {
@@ -102,28 +102,28 @@
 	justify-content: center;
 }
 
-.waterRightside img{
+.waterRightside img {
 	object-fit: cover;
 }
 
 .rightsideImage {
 	width: 600px;
 	height: 450px;
-    object-fit: cover;
+	object-fit: cover;
 }
 
 @media screen and (max-width: 700px) {
-	.container{
+	.container {
 		height: auto;
 	}
-	.getstarted{
+	.getstarted {
 		align-self: center;
 	}
 	.water {
 		flex-direction: column;
 	}
 
-	.waterLeftside{
+	.waterLeftside {
 		width: 100%;
 		gap: 20px;
 		margin: 10px 0;
@@ -131,8 +131,7 @@
 		background-repeat: no-repeat;
 		background-size: contain;
 	}
-	.waterRightside{
+	.waterRightside {
 		width: 100%;
 	}
-
 }
