+++ src/components/shared/Proof/Proof.module.css
@@ -4,6 +4,7 @@
 	margin: 10px 0;
 	max-width: 1000px;
 	background: none;
+	border-radius: 10px;
 }
 
 .Proof p {
